requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
openai>=1.12.0
google-generativeai>=0.3.0
tiktoken>=0.7.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.0.0
pdfplumber>=0.11.0
python-docx>=1.0.0
pypdf>=3.17.0
numpy>=1.24.0
pyyaml>=6.0.1
pillow>=10.0.0
pytesseract>=0.3.10
